+++ src/black/files.py
@@ -22,6 +22,7 @@ import tomli
 
 from black.output import err
 from black.report import Report
+from black.handle_ipynb_magics import jupyter_dependencies_are_installed
 
 if TYPE_CHECKING:
     import colorama  # noqa: F401
@@ -224,11 +225,10 @@ def gen_python_files(
             )
 
         elif child.is_file():
-            if child.suffix == ".ipynb":
-                from black import jupyter_dependencies_are_installed
-
-                if not jupyter_dependencies_are_installed(verbose=verbose, quiet=quiet):
-                    continue
+            if child.suffix == ".ipynb" and not jupyter_dependencies_are_installed(
+                verbose=verbose, quiet=quiet
+            ):
+                continue
             include_match = include.search(normalized_path) if include else True
             if include_match:
                 yield child
